requirements.txt
pypdf2
streamlit
python-dotenv
faiss-cpu
langchain 
openai
huggingface_hub
InstructorEmbedding
sentence_transformers
langchain-community